+++ src/lib/pages/publish-module/utils.ts
@@ -1,9 +1,10 @@
 import type { DecodeModuleQueryResponse } from "lib/services/moduleService";
-import type { Option } from "lib/types";
+import type { HumanAddr, Option } from "lib/types";
 import { UpgradePolicy } from "lib/types";
-import { truncate } from "lib/utils";
+import { bech32AddressToHex, truncate, unpadHexAddress } from "lib/utils";
 
-import type { FileArrayFields, PublishStatus } from "./formConstants";
+import type { Module, PublishStatus } from "./formConstants";
+import { publishStatusDefault } from "./formConstants";
 
 const priority = Object.keys(UpgradePolicy);
 
@@ -17,15 +18,22 @@ export const statusResolver = ({
   fields,
   index,
   policy,
+  address,
 }: {
   data: Option<DecodeModuleQueryResponse>;
-  fields: FileArrayFields;
+  fields: Module[];
   index: number;
   policy: UpgradePolicy;
+  address: Option<HumanAddr>;
 }): PublishStatus => {
-  if (!data) return { status: "init", text: "" };
+  if (!data) return publishStatusDefault;
 
-  const { abi, validPublisher, currentPolicy, modulePath } = data;
+  const { abi, currentPolicy, modulePath } = data;
+
+  const validPublisher = address
+    ? unpadHexAddress(bech32AddressToHex(address as HumanAddr)) === abi.address
+    : false;
+
   const priorUpload = fields
     .slice(0, index)
     .findLast((field) => field.decodeRes?.modulePath === modulePath);
@@ -85,5 +93,5 @@ export const statusResolver = ({
       text: `The file will be uploaded to republish module “${abi.name}” in your address.`,
     };
   }
-  return { status: "init", text: "" };
+  return publishStatusDefault;
 };